+++ ServerFrontEnd/src/app/components/home/home.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { AppSettings } from './../../appSettings';
 import { AuthenticationService } from 'src/app/services/authentication/authentication.service';
 import { Router } from '@angular/router';
+import { ArticlesService } from 'src/app/services/controllers/articles-controller.service';
 
 @Component({
   selector: 'home',
@@ -37,4 +38,7 @@ export class HomeComponent implements OnInit {
     return (this.router.url == button);
   }
 
+  onEnter(value) {
+    this.router.navigate(["/articles"], {queryParams: {value: value}});
+  }
 }