+++ ServerFrontEnd/src/app/services/controllers/articles-controller.service.ts
@@ -3,6 +3,7 @@ import { ArticleModel } from 'src/app/models/article.model';
 import { HttpClient } from '@angular/common/http';
 import { Injectable, Injector } from '@angular/core';
 import { AppRoutesService } from '../routes/app-routes.service';
+import { Observable } from 'rxjs';
 
 @Injectable({
     providedIn: 'root'
@@ -22,6 +23,10 @@ export class ArticlesService extends TemplateControllerService<ArticleModel> {
 
     protected getApiUrlObject() {
         return this.appRoutes.apiArticleEndPoint;
+    }
+
+    searchArticles(articleName: string): Observable<any> {
+        return this.http.get(`${this.appRoutes.apiArticleEndPointSearch}/${articleName}`);
     }
     
 }